Added vars for ambigous placement of updater script, and alias to update

+++ updaterc.sh
@@ -1,23 +1,31 @@
-#!/bin/zsh
+#!/bin/sh
 
-# Set $EDZSHDIR equal to where you ran the update script
-EDZSHDIR="$(dirname "$0")"
-cd $EDZSHDIR
+# Set variable for where the update script has been placed
+SCRIPT=$(readlink -f $0)
+SCRIPTDIR=`dirname $SCRIPT`
+
+# Verify the installation of ZSH, if not there install it (only on systems with apt currently
+if [ ! -f /usr/bin/zsh ] && [ -f /usr/bin/apt ]; then sudo apt update && sudo apt install zsh -y; fi
 
 # Pull the latest changes from the source
 git pull
-chmod +x $EDZSHDIR/updaterc.sh
 
-# Remove the original config
-rm ~/.zshrc
+# Remove the original config if it exists
+if [ -f ~/.zshrc ]; then rm ~/.zshrc; fi
 
 # Copy the new config to your home
-cp $EDZSHDIR/.zshrc ~/.zshrc
+cp $SCRIPTDIR/.zshrc ~/.zshrc
 
 # Add users' custom private configs if they exist
-if [ -f $EDZSHDIR/local_config ]; then cat $EDZSHDIR/local_config >> ~/.zshrc; fi
+if [ -f $SCRIPTDIR/local_config ]; then cat $SCRIPTDIR/local_config >> ~/.zshrc; fi
+
+# Add updater script location to .zshrc
+echo alias updaterc="'"$SCRIPT"'" >> ~/.zshrc
 
 echo Your .zshrc file has been updated!
+
+# Clean that screen!
+clear
 
 # Re-launch zsh to refresh the config
 /usr/bin/zsh
